Syncany-cli updated
