+++ syncany-cli/src/test/integration/java/org/syncany/tests/integration/cli/WatchCommandTest.java
@@ -15,7 +15,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package org.syncany.tests.cli;
+package org.syncany.tests.integration.cli;
 
 import static org.junit.Assert.assertTrue;
 import static org.syncany.tests.util.TestAssertUtil.assertFileEquals;
