+++ syncany-cli/src/test/unit/java/org/syncany/tests/util/SplitOutputStream.java
