+++ syncany-cli/src/test/unit/java/org/syncany/tests/util/TestCliUtil.java
